Move buffer related exceptions to sardanabuffer module
Late and Early value exceptions are relaed to the buffer. Move them to be
located close together.
